Lib: PowerSwitch: Add generic, static, and fix MOSFET
Add a generic PowerSwitch
Add PowerSwitchStatic, which just translates an ElectricLogic to an
ElectricPower
Fix the pull direction of the resistor on PowerSwitchMOSFET
